fix(dev): Fix errors in ADC and increase test-coverage

+++ denspp/offline/analog/common_func.py
@@ -47,7 +47,7 @@ class CommonDigitalFunctions:
         """
         xout = deepcopy(xin).astype(float)
         np.clip(xout, a_min=self._digital_border[0], a_max=self._digital_border[1], out=xout)
-        return xout.astype(np.int32)
+        return xout
 
     def _quantize_fxp(self, xin: np.ndarray) -> np.ndarray:
         """Do signed quantization of input with full precision
